+++ ZoneScouter/ZoneScouter.cs
@@ -12,11 +12,10 @@ using static PluginConfig;
 public sealed class ZoneScouter : BaseUnityPlugin {
   public const string PluginGuid = "redseiko.valheim.zonescouter";
   public const string PluginName = "ZoneScouter";
-  public const string PluginVersion = "1.6.0";
+  public const string PluginVersion = "1.7.0";
 
   void Awake() {
     BindConfig(Config);
-    ZoneSystemUtils.SetupUtils();
 
     Harmony.CreateAndPatchAll(Assembly.GetExecutingAssembly(), harmonyInstanceId: PluginGuid);
   }